Provide description for UI scale setting
